requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
pandas==2.0.3
numpy==1.24.4
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
